Moved CLI class to Tools folder
